+++ extensions/data-transfer/portability-data-transfer-google/src/main/java/org/dataportabilityproject/datatransfer/google/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Implements importers and exporters for Google services.
  */
-package org.dataportabilityproject.transfer.microsoft;
+package org.dataportabilityproject.datatransfer.google;
