+++ extensions/data-transfer/portability-data-transfer-microsoft/src/main/java/org/dataportabilityproject/datatransfer/microsoft/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Implements importers and exporters for Google services.
  */
-package org.dataportabilityproject.transfer.google;
+package org.dataportabilityproject.datatransfer.microsoft;
